feat(UI): Add PiP function to controls and allow configure documentPictureInPicture
