+++ ui/externs/ui.js
@@ -87,7 +87,8 @@ shaka.extern.UIVolumeBarColors;
  *   enableTooltips: boolean,
  *   keyboardSeekDistance: number,
  *   keyboardLargeSeekDistance: number,
- *   fullScreenElement: HTMLElement
+ *   fullScreenElement: HTMLElement,
+ *   preferDocumentPictureInPicture: boolean
  * }}
  *
  * @property {!Array.<string>} controlPanelElements
@@ -193,6 +194,10 @@ shaka.extern.UIVolumeBarColors;
  * @property {HTMLElement} fullScreenElement
  *   DOM element on which fullscreen will be done.
  *   Defaults to Shaka Player Container.
+ * @property {boolean} preferDocumentPictureInPicture
+ *   Indicates whether the Document Picture in Picture API is preferred or the
+ *   Video Element Picture in Picture API is preferred.
+ *   Defaults to true.
  * @exportDoc
  */
 shaka.extern.UIConfiguration;
