+++ ui/ui.js
@@ -245,6 +245,7 @@ shaka.ui.Overlay = class {
       keyboardSeekDistance: 5,
       keyboardLargeSeekDistance: 60,
       fullScreenElement: this.videoContainer_,
+      preferDocumentPictureInPicture: true,
     };
 
     // Check AirPlay support
